+++ wecom-objects/src/main/java/cn/felord/domain/approval/TmpControl.java
@@ -15,7 +15,7 @@
 
 package cn.felord.domain.approval;
 
-import cn.felord.convert.TmpControlDeserializer;
+import cn.felord.json.TmpControlDeserializer;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
