+++ wecom-objects/src/main/java/cn/felord/xml/XStreamXmlReader.java
@@ -15,8 +15,8 @@
 
 package cn.felord.xml;
 
-import cn.felord.convert.UnixInstantConverter;
 import cn.felord.domain.callback.Xml;
+import cn.felord.xml.convert.UnixInstantConverter;
 import com.thoughtworks.xstream.XStream;
 import com.thoughtworks.xstream.io.xml.DomDriver;
 import com.thoughtworks.xstream.io.xml.XmlFriendlyNameCoder;
