+++ wecom-objects/src/main/java/cn/felord/xml/convert/CallbackEventConverter.java
@@ -12,39 +12,42 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
+package cn.felord.xml.convert;
 
-package cn.felord.convert;
-
-import com.thoughtworks.xstream.converters.ConversionException;
+import cn.felord.enumeration.CallbackEvent;
 import com.thoughtworks.xstream.converters.basic.AbstractSingleValueConverter;
 
-import java.time.Instant;
-import java.time.format.DateTimeParseException;
+import java.util.Arrays;
+import java.util.Objects;
+
 
 /**
- * @author felord
- * @since 2021/10/12 22:15
+ * CallbackEvent
+ *
+ * @author dax
+ * @since 1.1.1
  */
-public class UnixInstantConverter extends AbstractSingleValueConverter {
+public class CallbackEventConverter extends AbstractSingleValueConverter {
+
     @Override
-    public boolean canConvert(final Class type) {
-        return Instant.class == type;
+    public boolean canConvert(Class type) {
+        return type != null && CallbackEvent.class.isAssignableFrom(type);
     }
 
     @Override
     public String toString(Object obj) {
-        Instant instant = (Instant) obj;
-        return String.valueOf(instant.toEpochMilli());
+        if (Objects.isNull(obj)) {
+            return null;
+        }
+        CallbackEvent eventEnum = (CallbackEvent) obj;
+        return eventEnum.type();
     }
 
     @Override
-    public Instant fromString(final String str) {
-        try {
-            return Instant.ofEpochSecond(Long.parseLong(str));
-        } catch (final DateTimeParseException ex) {
-            final ConversionException exception = new ConversionException("Cannot parse value as instant", ex);
-            exception.add("value", str);
-            throw exception;
-        }
+    public Object fromString(String str) {
+        return Arrays.stream(CallbackEvent.values())
+                .filter(callbackEvent -> Objects.equals(callbackEvent.type(), str))
+                .findFirst()
+                .orElse(null);
     }
 }
