+++ wecom-objects/src/main/java/cn/felord/xml/convert/CallbackNumberEnum.java
@@ -13,7 +13,7 @@
  *  limitations under the License.
  */
 
-package cn.felord.convert;
+package cn.felord.xml.convert;
 
 /**
  * @author dax
